+++ CHANGELOG.md
@@ -6,6 +6,8 @@ project uses [semantic versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-09-21
+
 ### Added
 
 - **A controller-layout button beside Desktop and Steam Big Picture.** Each